streamlit/requirements.txt
streamlit
sqlalchemy
psycopg2-binary
python-dotenv
pandas
plotly
folium
streamlit-folium
numpy
scikit-learn
altair
